requirements.txt
fastapi==0.109.0
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.31.0
requests==2.31.0
openai>=1.3.0
python-dotenv==1.0.0
google-genai==0.3.0
Pillow==10.2.0
python-multipart==0.0.9
